MultiAffine.jl

This package consists of an implementation of

the MultiAffine group

an action on vector spaces

MultiAffine Group

MultiAffine.MultiAffineGroup — Method

MultiAffineGroup(G, k=1)

An affine group modelling matrices of the form

\[χ = \begin{bmatrix} \mathbf{1} & \mathbf{0} \\ X & g \end{bmatrix}\]

where $g$ is a matrix element of the group $G$, represented in dimension $n$, and $X$ is a $n × k$ matrix. If we denote such an element by $[X,g]$, the multiplication law is

\[[X,g] [X',g'] = [X+gX';gg']\]

MultiAffine.MultiDisplacement — Type

MultiDisplacement(n, k=1)

A special case of the MultiAffineGroup group, where the underlying group is the special orthogonal group $SO(n)$. This is just a convenient alias

MultiDisplacement(n, k=1) = MultiAffineGroup(SpecialOrthogonal(n), k)

When $k=1$ (the default), this is the Special Euclidean Group.

MultiAffine Action

MultiAffine.MultiAffineAction — Type

MultiAffineAction(
  group::MultiAffineGroup,
  selector::AbstractArray,
  conv::ActionDirection=LeftAction()
  )

Given a fixed vector $S$ of size $k$ (the selector) (or a matrix of size $k×m$), this defines an action of the element $[X;R]$ of the MultiAffineGroup group (so $X$ is a $n×k$ matrix and $R$ is an element of a matrix group) on the vector $p$ of size $n$ (or the matrix $p$ of size $n×m$). The action is defined by

\[[X;R]⋅p := XS+Rp\]

MultiAffine.MultiAffineAction — Type

MultiAffineAction(
  group::MultiAffineGroup{<:Any, <:Any, 1},
  conv::ActionDirection=LeftAction()
  )

The standard affine action, obtained with a selector equal to $[1]$.